+++ src/main/java/msg/team1/Hi/domain/auth/entity/BlackList.java
@@ -1,0 +1,21 @@
+package msg.team1.Hi.domain.auth.entity;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.springframework.data.annotation.Id;
+import org.springframework.data.redis.core.RedisHash;
+import org.springframework.data.redis.core.TimeToLive;
+
+@Getter @Builder
+@NoArgsConstructor
+@AllArgsConstructor
+@RedisHash(value = "blackList")
+public class BlackList {
+    @Id
+    private String accessToken;
+    private String email;
+    @TimeToLive
+    private Long timeToLive;
+}
