+++ src/main/java/msg/team1/Hi/domain/auth/exception/RefreshTokenNotFoundException.java
@@ -1,0 +1,15 @@
+package msg.team1.Hi.domain.auth.exception;
+
+import lombok.Getter;
+import msg.team1.Hi.global.exception.ErrorCode;
+
+@Getter
+public class RefreshTokenNotFoundException extends RuntimeException {
+    private final ErrorCode errorCode;
+
+    public RefreshTokenNotFoundException(String message) {
+        super(message);
+        this.errorCode = ErrorCode.REFRESH_TOKEN_NOT_FOUND;
+    }
+
+}
