+++ src/main/java/msg/team1/Hi/domain/auth/presentation/dto/request/LoginRequest.java
@@ -1,0 +1,20 @@
+package msg.team1.Hi.domain.auth.presentation.dto.request;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import javax.validation.constraints.NotEmpty;
+
+@Getter
+@AllArgsConstructor
+@NoArgsConstructor
+public class LoginRequest {
+
+    @NotEmpty
+    private String email;
+
+    @NotEmpty
+    private String password;
+
+}
