+++ src/main/java/msg/team1/Hi/domain/email/service/EmailService.java
@@ -1,6 +1,6 @@
 package msg.team1.Hi.domain.email.service;
 
-import msg.team1.Hi.domain.email.dto.request.EmailSentDto;
+import msg.team1.Hi.domain.email.presentation.dto.request.EmailSentDto;
 
 public interface EmailService {
     void sendEmail(EmailSentDto emailSentDto);
