+++ src/main/java/msg/team1/Hi/domain/home_base/service/HomeBaseService.java
@@ -1,6 +1,6 @@
 package msg.team1.Hi.domain.home_base.service;
 
-import msg.team1.Hi.domain.home_base.dto.request.ReserveHomeBaseRequest;
+import msg.team1.Hi.domain.home_base.presentation.dto.request.ReserveHomeBaseRequest;
 
 public interface HomeBaseService {
     void reserveHomeBase(ReserveHomeBaseRequest request);
