--- src/main/java/msg/team1/Hi/domain/member/dto/response/MemberLoginResponse.java
@@ -1,18 +1,0 @@
-package msg.team1.Hi.domain.member.dto.response;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-import lombok.*;
-
-import java.time.ZonedDateTime;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class MemberLoginResponse {
-    private String accessToken;
-    private String refreshToken;
-
-    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ss")
-    private ZonedDateTime expiredAt;
-}
