--- src/main/java/msg/team1/Hi/domain/member/dto/response/NewTokenResponse.java
@@ -1,22 +1,0 @@
-package msg.team1.Hi.domain.member.dto.response;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.time.ZonedDateTime;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class NewTokenResponse {
-
-    private String accessToken;
-    private String refreshToken;
-
-    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ss")
-    private ZonedDateTime expiredAt;
-}
