--- src/main/java/msg/team1/Hi/domain/notice/dto/response/GetIdNoticeResponse.java
@@ -1,24 +1,0 @@
-package msg.team1.Hi.domain.notice.dto.response;
-
-import lombok.*;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-
-import java.time.LocalDateTime;
-
-@Getter @Builder
-@NoArgsConstructor(access = AccessLevel.PROTECTED) @AllArgsConstructor
-public class GetIdNoticeResponse {
-
-    private Integer noticeId;
-    private String title;
-    private String content;
-    private String name;
-
-    @CreatedDate
-    private LocalDateTime createdDate;
-
-    @LastModifiedDate
-    private LocalDateTime modifiedDate;
-
-}
