+++ src/main/java/msg/team1/Hi/global/exception/handler/GlobalExceptionHandler.java
@@ -6,10 +6,10 @@ import msg.team1.Hi.domain.email.exception.ManyRequestEmailAuthException;
 import msg.team1.Hi.domain.email.exception.MisMatchAuthCodeException;
 import msg.team1.Hi.domain.email.exception.NotVerifyEmailException;
 import msg.team1.Hi.domain.home_base.exception.ReservedHomeBaseException;
-import msg.team1.Hi.domain.member.exception.ExistEmailException;
-import msg.team1.Hi.domain.member.exception.MemberNotFoundException;
+import msg.team1.Hi.domain.auth.exception.ExistEmailException;
+import msg.team1.Hi.domain.auth.exception.MemberNotFoundException;
 import msg.team1.Hi.domain.member.exception.MisMatchPasswordException;
-import msg.team1.Hi.domain.member.exception.RefreshTokenNotFoundException;
+import msg.team1.Hi.domain.auth.exception.RefreshTokenNotFoundException;
 import msg.team1.Hi.domain.notice.exception.NoticeNotFoundException;
 import msg.team1.Hi.global.exception.ErrorMessage;
 import msg.team1.Hi.global.exception.collection.TokenExpirationException;
